various module import renames
