+++ src/Language/Haskell/Liquid/DiffCheck.hs
@@ -42,7 +42,7 @@ import qualified  Data.HashMap.Strict           as M
 import qualified  Data.List                     as L
 import            System.Directory                (copyFile, doesFileExist)
 import            Language.Fixpoint.Types         (FixResult (..), Located (..))
-import            Language.Fixpoint.Files
+import            Language.Fixpoint.Utils.Files
 import            Language.Haskell.Liquid.Types   (SpecType, GhcSpec (..), AnnInfo (..), DataConP (..), Error, TError (..), Output (..))
 import            Language.Haskell.Liquid.Misc    (mkGraph)
 import            Language.Haskell.Liquid.GhcMisc
