+++ src/Language/Haskell/Liquid/Parse.hs
@@ -42,7 +42,7 @@ import Language.Haskell.Liquid.Variance
 import Language.Haskell.Liquid.Bounds
 
 import qualified Language.Haskell.Liquid.Measure as Measure
-import Language.Fixpoint.Names (symbolString, listConName, hpropConName, propConName, tupConName, headSym)
+import Language.Fixpoint.Types.Names (symbolString, listConName, hpropConName, propConName, tupConName, headSym)
 import Language.Fixpoint.Misc (safeLast, errorstar)
 import Language.Fixpoint.Parse hiding (angles, refBindP, refP, refDefP)
